refactor: extract cli.py dump and compare command bodies to helper modules
The two largest CodeFactor complex-method findings. Pure refactors —
CLI behavior, the full exit-code matrix, output, and error messages
unchanged:

- dump: the collect-context and compile-context resolution blocks move to
  cli_dump_helpers.py; the click command stays a thin option-parsing
  wrapper.
- compare: the ~417-line body moves to a new click-free
  cli_compare_helpers.run_compare(); the wrapper forwards parsed options
  verbatim. Verdict routing still goes through the Tier-2 service
  (service.compare_snapshots), never checker.compare (cli-contract clean).
  monkeypatch-by-name resolution preserved via cli-module attribute access.

cli.py 1988 → 1537 lines. Adds the by-design cli/cli_compare_helpers
sibling cycle to IMPORT_CYCLE_ALLOWLIST.
