+++ abicheck/cli_compare_helpers.py
@@ -1,0 +1,549 @@
+# Copyright 2026 Nikolay Petrov
+# SPDX-License-Identifier: Apache-2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Orchestration body for the ``compare`` command (size-split from cli.py).
+
+The click-decorated ``compare`` wrapper in :mod:`abicheck.cli` parses options and
+delegates to :func:`run_compare` here, keeping cli.py under the AI-readiness
+file-size cap. This is *not* the leaf helper module ``cli_helpers_compare`` (plain,
+cli-independent utilities): ``run_compare`` drives the full single-pair compare
+flow and reuses the option-parsing/render/exit helpers that still live in
+:mod:`abicheck.cli` (imported back below — the by-design sibling cycle, allow-listed
+in ``check_ai_readiness``). Verdict routing stays through the Tier-2 service
+(``service.compare_snapshots``), never a direct ``checker.compare`` call
+(cli-contract, ADR-037 D10.1).
+"""
+
+from __future__ import annotations
+
+from pathlib import Path
+
+import click
+
+from . import cli
+from .cli import (
+    _announce_exit_scheme,
+    _embed_inline_source_side,
+    _exit_with_severity_or_verdict,
+    _finalize_compare_result,
+    _load_probe_matrix_changes,
+    _log_debug_resolution,
+    _reject_application_operand,
+    _render_output,
+    _setup_verbosity,
+    _source_is_pack,
+    _warn_unused_set_flags,
+    _write_or_echo,
+)
+from .cli_audit import echo_pattern_modulations
+from .cli_dump_helpers import resolve_dump_depth
+from .cli_helpers_compare import (
+    _collect_force_public_symbols,
+    _resolve_per_side_options,
+    _warn_ignored_flags,
+)
+from .cli_options import resolve_compile_context
+from .cli_params import _load_suppression_and_policy
+from .cli_resolve import (
+    _reject_compile_context_for_set_inputs,
+    _reject_evidence_flags_for_set_inputs,
+    _resolve_compare_snapshots,
+    classify_compare_operand,
+)
+from .errors import AbicheckError
+
+
+def run_compare(
+    ctx: click.Context,
+    *,
+    old_input: Path, new_input: Path,
+    jobs: int, dso_only: bool, output_dir: Path | None,
+    fail_on_removed: bool,
+    debug_info1: Path | None, debug_info2: Path | None,
+    devel_pkg1: Path | None, devel_pkg2: Path | None,
+    include_private_dso: bool, keep_extracted: bool,
+    manifest_path: Path | None, bundle_system_providers: str,
+    bundle_cohorts: tuple[str, ...], no_bundle_analysis: bool,
+    headers: tuple[Path, ...], includes: tuple[Path, ...], lang: str,
+    header_backend: str,
+    gcc_path: str | None, gcc_prefix: str | None, gcc_options: str | None,
+    gcc_option_tokens: tuple[str, ...], sysroot: Path | None, nostdinc: bool,
+    old_header_backend: str | None, new_header_backend: str | None,
+    old_headers_only: tuple[Path, ...], new_headers_only: tuple[Path, ...],
+    old_includes_only: tuple[Path, ...], new_includes_only: tuple[Path, ...],
+    old_version: str, new_version: str,
+    fmt: str, demangle: bool | None, output: Path | None,
+    suppress: Path | None, strict_suppressions: bool, require_justification: bool,
+    policy: str, policy_file_path: Path | None,
+    pdb_path: Path | None, old_pdb_path: Path | None, new_pdb_path: Path | None,
+    dwarf_only: bool,
+    severity_preset: str | None,
+    severity_abi_breaking: str | None,
+    severity_potential_breaking: str | None,
+    severity_quality_issues: str | None,
+    severity_addition: str | None,
+    config: Path | None,
+    exit_code_scheme: str | None,
+    follow_deps: bool, search_paths: tuple[Path, ...], ld_library_path: str,
+    show_redundant: bool, show_only: str | None, stat: bool,
+    scope_public_headers: bool, collapse_versioned_symbols: bool, show_filtered: bool,
+    public_symbols: tuple[str, ...], public_symbols_list: Path | None,
+    post_manifest_path: Path | None,
+    report_mode: str, show_impact: bool,
+    recommend: bool,
+    debug_format_opt: str | None,
+    debug_format: str | None,
+    annotate: bool,
+    annotate_additions: bool,
+    debug_roots: tuple[Path, ...],
+    debug_roots_old: tuple[Path, ...],
+    debug_roots_new: tuple[Path, ...],
+    debuginfod: bool,
+    debuginfod_url: str | None,
+    pattern_verdicts: bool,
+    explain_patterns: bool,
+    surface_metrics: bool,
+    reconcile_build_context: bool,
+    env_matrix_path: Path | None,
+    verbose: bool,
+    old_build_info: Path | None = None, new_build_info: Path | None = None,
+    old_sources: Path | None = None, new_sources: Path | None = None,
+    depth: str | None = None, max_depth: bool = False,
+    probe_matrix_old: Path | None = None,
+    probe_matrix_new: Path | None = None,
+) -> None:
+    """Run the single-pair (or set fan-out) ``compare`` flow and exit accordingly."""
+    _setup_verbosity(verbose)
+
+    # ADR-037 D4: load the project config and merge CLI flags over it
+    # (precedence CLI > config > built-in default) *before* dispatch, so both the
+    # single-file and the directory/package fan-out paths share one resolution.
+    # Auto-discovered from the current directory upward, overridable with --config.
+    from .buildsource.inline import load_build_config
+    from .cli_helpers_compare import discover_project_config, resolve_compare_config
+
+    cfg_path = config if config is not None else discover_project_config()
+    try:
+        project_cfg = load_build_config(cfg_path) if cfg_path is not None else None
+    except ValueError as exc:
+        raise click.UsageError(str(exc)) from exc
+
+    def _cli_flag(name: str, value: bool) -> bool | None:
+        # Return the value only when it actually came from the command line, so a
+        # flag default (e.g. --scope-public-headers's True) doesn't mask config.
+        src = click.get_current_context().get_parameter_source(name)
+        return value if src == click.core.ParameterSource.COMMANDLINE else None
+
+    resolved_cfg = resolve_compare_config(
+        project_cfg,
+        cli_severity_preset=severity_preset,
+        cli_severity_abi_breaking=severity_abi_breaking,
+        cli_severity_potential_breaking=severity_potential_breaking,
+        cli_severity_quality_issues=severity_quality_issues,
+        cli_severity_addition=severity_addition,
+        cli_scope_public=_cli_flag("scope_public_headers", scope_public_headers),
+        cli_collapse_versioned_symbols=_cli_flag(
+            "collapse_versioned_symbols", collapse_versioned_symbols
+        ),
+        cli_public_symbols=public_symbols,
+        cli_strict_suppressions=_cli_flag("strict_suppressions", strict_suppressions),
+        cli_require_justification=_cli_flag("require_justification", require_justification),
+        cli_exit_code_scheme=exit_code_scheme,
+    )
+    sev_config = resolved_cfg.severity
+    scope_public_headers = resolved_cfg.scope_public
+    collapse_versioned_symbols = resolved_cfg.collapse_versioned_symbols
+    strict_suppressions = resolved_cfg.strict_suppressions
+    require_justification = resolved_cfg.require_justification
+
+    # ADR-037 D7: input-type dispatch. A directory or package operand fans out to
+    # a per-library comparison (absorbing `compare-release`); an application/PIE
+    # operand is not a library `compare` can pair (hint at `appcompat`). A single
+    # .so / snapshot / dump falls through to the normal one-pair path below. The
+    # resolved config (scope/suppression/severity) is forwarded so a set-input
+    # compare classifies the same way a single-pair one would (ADR-037 D4).
+    old_kind = classify_compare_operand(old_input)
+    new_kind = classify_compare_operand(new_input)
+    if old_kind == "app" or new_kind == "app":
+        _reject_application_operand(old_input, new_input, old_kind, new_kind)
+    if {old_kind, new_kind} & {"directory", "package"}:
+        # The per-library fan-out (`compare-release` backend) consumes the
+        # resolved scheme from config, but still has no public CLI support for
+        # an explicit --exit-code-scheme on set inputs. Reject the CLI flag
+        # loudly rather than silently ignore it (ADR-037 D12).
+        if exit_code_scheme is not None:
+            raise click.UsageError(
+                "--exit-code-scheme is not supported for directory/package "
+                "(release) comparisons: the per-library fan-out uses the legacy "
+                "verdict scheme, or severity-aware when severity is configured in "
+                ".abicheck.yml. Compare libraries individually for explicit "
+                "scheme control."
+            )
+        # --reconcile-build-context (ADR-039) is not threaded through the
+        # per-library release fan-out; reject it loudly rather than silently
+        # ignore it (ADR-037 D12). It applies to single-file / snapshot inputs.
+        if reconcile_build_context:
+            raise click.UsageError(
+                "--reconcile-build-context is not supported for directory/package "
+                "(release) comparisons; it applies to single-file / snapshot "
+                "inputs. Compare the libraries individually to use it."
+            )
+        # --env-matrix (ADR-020b) is likewise not threaded through the release
+        # fan-out; a silently ignored runtime floor would leave the default
+        # RISK verdicts in place while the user believes the contract applied.
+        if env_matrix_path is not None:
+            raise click.UsageError(
+                "--env-matrix is not supported for directory/package (release) "
+                "comparisons yet; it applies to single-file / snapshot inputs. "
+                "Compare the libraries individually to use it."
+            )
+        _reject_compile_context_for_set_inputs(ctx, project_cfg)
+        _reject_evidence_flags_for_set_inputs(ctx)
+        # Resolved through the ``cli`` module (not a by-name import) so a test that
+        # monkeypatches ``abicheck.cli._dispatch_release_compare`` before invoking
+        # ``compare`` is honoured — matching the pre-split resolution semantics.
+        cli._dispatch_release_compare(
+            ctx,
+            old_dir=old_input, new_dir=new_input,
+            headers=headers, includes=includes,
+            old_headers_only=old_headers_only, new_headers_only=new_headers_only,
+            old_includes_only=old_includes_only, new_includes_only=new_includes_only,
+            old_version=old_version, new_version=new_version, lang=lang,
+            fmt=fmt, output=output, output_dir=output_dir,
+            suppress=suppress, strict_suppressions=strict_suppressions,
+            require_justification=require_justification,
+            policy=policy, policy_file_path=policy_file_path,
+            dso_only=dso_only, jobs=jobs,
+            fail_on_removed=fail_on_removed,
+            debug_info1=debug_info1, debug_info2=debug_info2,
+            devel_pkg1=devel_pkg1, devel_pkg2=devel_pkg2,
+            include_private_dso=include_private_dso, keep_extracted=keep_extracted,
+            manifest_path=manifest_path,
+            bundle_system_providers=bundle_system_providers,
+            bundle_cohorts=bundle_cohorts, no_bundle_analysis=no_bundle_analysis,
+            scope_public_headers=scope_public_headers,
+            severity_preset=resolved_cfg.merged_severity_preset,
+            severity_abi_breaking=resolved_cfg.merged_severity_abi_breaking,
+            severity_potential_breaking=resolved_cfg.merged_severity_potential_breaking,
+            severity_quality_issues=resolved_cfg.merged_severity_quality_issues,
+            severity_addition=resolved_cfg.merged_severity_addition,
+            release_exit_code_scheme=resolved_cfg.exit_code_scheme,
+            probe_matrix_old=probe_matrix_old, probe_matrix_new=probe_matrix_new,
+            annotate=annotate, annotate_additions=annotate_additions,
+            verbose=verbose,
+        )
+        return
+    # Single-file/snapshot inputs: the set-only fan-out flags do not apply.
+    jobs_explicit = (
+        ctx.get_parameter_source("jobs") == click.core.ParameterSource.COMMANDLINE
+    )
+    _warn_unused_set_flags(
+        jobs_explicit=jobs_explicit, dso_only=dso_only, output_dir=output_dir
+    )
+
+    if annotate_additions and not annotate:
+        raise click.UsageError("--annotate-additions requires --annotate")
+
+    # Fold the unified --depth/--max dial into the internal collect mode
+    # (ADR-037 D5), the same way `dump` does. With no preset, compare reads at
+    # "off"; --depth binary suppresses the L2 header AST (symbols-only).
+    collect_mode = resolve_dump_depth(depth, max_depth, "off")
+    if depth == "binary":
+        headers, old_headers_only, new_headers_only = (), (), ()
+
+    # Reconcile the --debug-format selector with the legacy --btf/--ctf/--dwarf
+    # flags. The selector supersedes the legacy flags whenever it is given:
+    # an explicit "auto" returns to auto-detection (None) even if a legacy flag
+    # is also present; only when the selector is absent do the legacy flags apply.
+    if debug_format_opt is not None:
+        effective_debug_format = None if debug_format_opt.lower() == "auto" else debug_format_opt
+    else:
+        effective_debug_format = debug_format
+
+    # Tri-state --demangle: default ON for the text formats whose renderer
+    # post-processes symbols through demangle_text (markdown/review), OFF for
+    # machine formats (json/sarif/junit) and HTML — the HTML renderer emits
+    # symbols structurally and demangling its string would inject unescaped
+    # '<'/'>'/'&' from C++ names and corrupt the markup. Explicit flag wins.
+    if demangle is None:
+        demangle = fmt in {"markdown", "review"}
+
+    # --report-mode impact is sugar for "full" report with the impact table on.
+    if report_mode == "impact":
+        report_mode = "full"
+        show_impact = True
+
+    # ADR-037 D4: the precise S-axis lives in config (source.method). It sets the
+    # collection depth only when the user gave no explicit --depth/--max signal
+    # (CLI > config).
+    if (
+        resolved_cfg.source_method
+        and depth is None
+        and not max_depth
+    ):
+        from .buildsource.scan_levels import SourceMethod, method_to_collect_mode
+        try:
+            collect_mode = method_to_collect_mode(SourceMethod(resolved_cfg.source_method))
+        except ValueError:
+            raise click.UsageError(
+                f"source.method in .abicheck.yml is invalid: "
+                f"{resolved_cfg.source_method!r} (expected s0..s6 or auto)."
+            ) from None
+
+    # L2 header compile context (compare↔dump↔scan parity, ADR-037 D3): the one
+    # shared resolver folds the project's .abicheck.yml compile: block into the CLI
+    # cross-toolchain/frontend flags (CLI > config) and appends config include_dirs
+    # after the -I roots. It applies to both sides; the per-side --old/new-ast-frontend
+    # overrides still win for the frontend (threaded separately below). cfg_path is
+    # the same config compare resolves everything else from (explicit --config or the
+    # .abicheck.yml auto-discovered from cwd).
+    import dataclasses
+
+    compile_context, merged_includes = resolve_compile_context(
+        ctx,
+        gcc_path=gcc_path, gcc_prefix=gcc_prefix, gcc_options=gcc_options,
+        gcc_option_tokens=gcc_option_tokens, sysroot=sysroot, nostdinc=nostdinc,
+        header_backend=header_backend, includes=includes, build_config=cfg_path,
+    )
+    # The dirs the config appended past the CLI -I roots. These are documented as
+    # applying to *both* sides, so they must survive a per-side --old/new-include
+    # override (which replaces the both-sides -I for that side). Keep them separate
+    # and re-append after per-side resolution rather than folding into the shared
+    # tuple, else the overridden side would lose them (Codex review).
+    config_includes = tuple(merged_includes[len(includes):])
+    # The merged frontend flows to both sides through the explicit header_backend
+    # (so --old/new-ast-frontend can still override per side); neutralize the
+    # frontend on the threaded context so run_dump's `compile.frontend` does NOT
+    # outrank that per-side header_backend (it only carries the --gcc-*/--sysroot/
+    # --nostdinc knobs for both sides).
+    header_backend = compile_context.frontend
+    side_compile_context = dataclasses.replace(compile_context, frontend="auto")
+
+    old_h, new_h, old_inc, new_inc = _resolve_per_side_options(
+        headers, includes, old_headers_only, new_headers_only,
+        old_includes_only, new_includes_only,
+    )
+    if config_includes:
+        old_inc = list(old_inc) + list(config_includes)
+        new_inc = list(new_inc) + list(config_includes)
+
+    # Inline source-tree collection (deep-compare folded into compare): when a
+    # side's --old/new-sources points at a raw checkout, or --old/new-build-info
+    # at a raw build dir / compile_commands.json (not a `collect` pack), dump that
+    # side at --depth so its L3-L5 facts ride embedded in the snapshot, the way
+    # the standalone deep-compare command used to. Pre-built packs fall through
+    # unchanged to prepare_embedded_build_source below.
+    def _raw_evidence(p: Path | None) -> bool:
+        return p is not None and not _source_is_pack(p)
+
+    if (
+        _raw_evidence(old_sources)
+        or _raw_evidence(new_sources)
+        or _raw_evidence(old_build_info)
+        or _raw_evidence(new_build_info)
+    ):
+        import shutil
+        import tempfile
+
+        # CLI-over-config explicitness read from compare's *real* ctx (where
+        # --ast-frontend/--nostdinc are genuine COMMANDLINE params); the inline
+        # dump runs under ctx.invoke where that signal is lost, so we compute it
+        # here and thread it through (Codex review). A per-side --old/new-ast-frontend
+        # is itself an explicit frontend for that side.
+        _nostdinc_explicit = (
+            ctx.get_parameter_source("nostdinc")
+            == click.core.ParameterSource.COMMANDLINE
+        )
+        _frontend_explicit = (
+            ctx.get_parameter_source("header_backend")
+            == click.core.ParameterSource.COMMANDLINE
+        )
+
+        _src_tmp = tempfile.mkdtemp(prefix="abicheck-compare-src-")
+        # Cleanup on context teardown so the temp dir never leaks, even if an
+        # inline dump or _resolve_compare_snapshots raises before we return.
+        ctx.call_on_close(lambda: shutil.rmtree(_src_tmp, ignore_errors=True))
+        old_input, old_sources, old_build_info = _embed_inline_source_side(
+            ctx, input_path=old_input, sources=old_sources,
+            headers=old_h, includes=old_inc, version=old_version, lang=lang,
+            header_backend=old_header_backend or header_backend,
+            compile_context=compile_context,
+            frontend_explicit=_frontend_explicit or old_header_backend is not None,
+            # A nostdinc already resolved True (from --config) must survive the
+            # tree-config merge even when the tree omits it (Codex review); False
+            # is the default and indistinguishable from "unset", so only True needs
+            # preserving.
+            nostdinc_explicit=_nostdinc_explicit or compile_context.nostdinc,
+            build_info=old_build_info,
+            follow_deps=follow_deps, search_paths=search_paths,
+            ld_library_path=ld_library_path,
+            dwarf_only=dwarf_only, debug_format=effective_debug_format,
+            pdb_path=old_pdb_path or pdb_path,
+            collect_mode=collect_mode, out_dir=Path(_src_tmp), label="old",
+        )
+        new_input, new_sources, new_build_info = _embed_inline_source_side(
+            ctx, input_path=new_input, sources=new_sources,
+            headers=new_h, includes=new_inc, version=new_version, lang=lang,
+            header_backend=new_header_backend or header_backend,
+            compile_context=compile_context,
+            frontend_explicit=_frontend_explicit or new_header_backend is not None,
+            nostdinc_explicit=_nostdinc_explicit or compile_context.nostdinc,
+            build_info=new_build_info,
+            follow_deps=follow_deps, search_paths=search_paths,
+            ld_library_path=ld_library_path,
+            dwarf_only=dwarf_only, debug_format=effective_debug_format,
+            pdb_path=new_pdb_path or pdb_path,
+            collect_mode=collect_mode, out_dir=Path(_src_tmp), label="new",
+        )
+
+    # Follow GNU ld linker scripts up front so the resolved DSO (not the text
+    # script) drives format detection, metadata, and dependency analysis.
+    # Through the ``cli`` module so a monkeypatch on ``abicheck.cli._normalize_binary_input``
+    # is honoured (pre-split resolution semantics); the name is re-exported there.
+    old_input, old_fmt = cli._normalize_binary_input(old_input)
+    new_input, new_fmt = cli._normalize_binary_input(new_input)
+    # --debug-format / legacy --btf/--ctf/--dwarf force an ELF debug format and
+    # are silently ignored by the PE/Mach-O dump paths. Reject them up front for
+    # non-ELF binary inputs (mirrors dump_cmd) so the flag is never accepted but
+    # ignored. JSON-snapshot / dump inputs have *_fmt == None and are unaffected.
+    if effective_debug_format is not None:
+        for side, bfmt in (("old", old_fmt), ("new", new_fmt)):
+            if bfmt in ("pe", "macho"):
+                raise click.BadParameter(
+                    f"--debug-format {effective_debug_format} is only supported "
+                    f"for ELF binaries, but the {side} input is {bfmt.upper()}."
+                )
+    _warn_ignored_flags(
+        old_fmt is not None, new_fmt is not None,
+        headers, includes,
+        old_headers_only, new_headers_only,
+        old_includes_only, new_includes_only,
+    )
+
+    # Resolve per-side debug roots: --debug-root1 overrides --debug-root for old, etc.
+    resolved_old_debug = list(debug_roots_old) if debug_roots_old else list(debug_roots)
+    resolved_new_debug = list(debug_roots_new) if debug_roots_new else list(debug_roots)
+    _log_debug_resolution(
+        old_input, new_input,
+        resolved_old_debug, resolved_new_debug,
+        debuginfod=debuginfod, debuginfod_url=debuginfod_url,
+    )
+
+    old, new = _resolve_compare_snapshots(
+        old_input, new_input, old_fmt, new_fmt,
+        old_h, new_h, old_inc, new_inc,
+        old_version, new_version, lang,
+        pdb_path, old_pdb_path, new_pdb_path,
+        dwarf_only, effective_debug_format,
+        follow_deps, search_paths, ld_library_path,
+        header_backend=header_backend,
+        old_header_backend=old_header_backend,
+        new_header_backend=new_header_backend,
+        compile_context=side_compile_context,
+    )
+
+    suppression, pf = _load_suppression_and_policy(
+        suppress, policy, policy_file_path,
+        strict_suppressions=strict_suppressions,
+        require_justification=require_justification,
+    )
+
+    force_public = _collect_force_public_symbols(
+        resolved_cfg.public_symbols, public_symbols_list
+    )
+    if force_public and not scope_public_headers:
+        click.echo(
+            "Warning: --public-symbol/--public-symbols-list only take effect with "
+            "--scope-public-headers; ignoring the widening overlay.",
+            err=True,
+        )
+
+    extra_changes = _load_probe_matrix_changes(probe_matrix_old, probe_matrix_new)
+
+    # Build-info + source facts (ADR-028/033): the helper times inline diffing
+    # for the D6/D9 metrics and returns coverage/metrics to attach post-compare.
+    from .cli_buildsource import attach_evidence_metrics, prepare_embedded_build_source
+    extra_changes, layer_coverage_rows, evidence_metrics, _ev_changes = (
+        prepare_embedded_build_source(
+            old, new, collect_mode, extra_changes,
+            old_build_info, new_build_info, old_sources, new_sources,
+            policy_file=pf,
+        )
+    )
+
+    # --post-manifest: scope the comparison to the POST manifest's committed
+    # `pp_*`/ufunc-loop surface. The manifest *is* the authoritative public
+    # surface, so this drives FilterNonPublicSurface directly (no header
+    # provenance needed) — private __pp_* kernel churn is demoted.
+    post_manifest_allowlist: set[str] | None = None
+    if post_manifest_path is not None:
+        from .post_manifest import contract_scope_allowlist, load_manifest
+
+        try:
+            manifest = load_manifest(post_manifest_path)
+        except (ValueError, OSError) as exc:
+            raise click.UsageError(f"--post-manifest {post_manifest_path}: {exc}") from exc
+        # Union with the binaries' committed (pp_*) exports so a *removed* wrapper
+        # — absent from a new manifest — stays in-surface instead of being
+        # silently demoted (its symbol still lives in the old snapshot).
+        post_manifest_allowlist = contract_scope_allowlist(manifest, old, new)
+
+    apply_patterns = pattern_verdicts or explain_patterns  # --explain implies on
+    from .service import compare_snapshots, load_env_matrix
+    try:
+        env_matrix = load_env_matrix(env_matrix_path)
+    except AbicheckError as exc:
+        raise click.UsageError(str(exc)) from exc
+    result = compare_snapshots(
+        old, new, suppression=suppression, policy=policy, policy_file=pf,
+        env_matrix=env_matrix,
+        scope_to_public_surface=scope_public_headers,
+        force_public_symbols=force_public,
+        extra_changes=extra_changes,
+        pattern_verdicts=apply_patterns,
+        surface_metrics=surface_metrics,
+        collapse_versioned_symbols=collapse_versioned_symbols,
+        public_surface_allowlist=post_manifest_allowlist,
+        reconcile_build_context=reconcile_build_context,
+    )
+    if layer_coverage_rows:
+        result.layer_coverage = layer_coverage_rows
+    # Pass all injected findings (probe-matrix + evidence) so artifact-backed
+    # excludes them — none come from L0-L2 diffing.
+    attach_evidence_metrics(result, evidence_metrics, extra_changes or [])
+
+    if explain_patterns:
+        echo_pattern_modulations(result)
+
+    _finalize_compare_result(
+        result, old_input, new_input,
+        show_redundant=show_redundant, show_filtered=show_filtered,
+        annotate=annotate, annotate_additions=annotate_additions,
+    )
+
+    text = _render_output(
+        fmt, result, old, new,
+        follow_deps=follow_deps,
+        show_only=show_only, report_mode=report_mode,
+        show_impact=show_impact, stat=stat,
+        severity_config=sev_config if resolved_cfg.exit_code_scheme == "severity" else None,
+        show_recommendation=recommend,
+        demangle=demangle,
+    )
+
+    _write_or_echo(output, text)
+
+    _announce_exit_scheme(resolved_cfg.exit_code_scheme, fmt=fmt, stat=stat)
+    _exit_with_severity_or_verdict(result, sev_config, resolved_cfg.exit_code_scheme)
